create src folder
