+++ tsconfig.json
@@ -60,5 +60,5 @@
     "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */,
     "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */
   },
-  "exclude": ["node_modules", "**/*.spec.ts", "build", "coverage", "dev-commands"]
+  "exclude": ["node_modules", "**/**/*.spec.ts", "build", "coverage", "dev-commands"]
 }
